Fix issue with coupon codes being removed at the checkout
When the billing address "use_for_shipping" flag was checked, editing
the billing address would wipe the coupon code. This was caused by the
item cache fields being saved on the billing address, being copied into
the shipping address, which would result in the coupon code not being
set during the collection of totals.

To resolve the issue we unset the item cache fields on the address
object, which allows the coupon code to be set correctly after editing
the billing address.